chore: remove common dir
