+++ packages/kit-bg/src/dbs/simple/entity/SimpleDbEntityAddressBook.ts
@@ -1,4 +1,4 @@
-import type { IAddressItem } from '@onekeyhq/kit/src/common/components/AddressBook/type';
+import type { IAddressItem } from '@onekeyhq/kit/src/views/AddressBook/type';
 
 import { SimpleDbEntityBase } from './SimpleDbEntityBase';
