--- packages/kit/src/common/hooks/useAddressBook.ts
@@ -1,1 +1,0 @@
-export * from '../components/AddressBook/hooks/useAddressBook';
