+++ packages/kit/src/routes/Modal/router.tsx
@@ -2,8 +2,8 @@ import type { IModalRootNavigatorConfig } from '@onekeyhq/components/src/layouts
 import { ModalSettingStack } from '@onekeyhq/kit/src/views/Setting/router';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
-import { ModalAddressBookRouter } from '../../common/components/AddressBook/router';
 import { AccountManagerStacks } from '../../views/AccountManagerStacks/router';
+import { ModalAddressBookRouter } from '../../views/AddressBook/router';
 import { AssetSelectorRouter } from '../../views/AssetSelector/router';
 import { ChainSelectorRouter } from '../../views/ChainSelector/router';
 import { DAppConnectionRouter } from '../../views/DAppConnection/router';
