+++ packages/kit/src/views/AddressBook/router/index.tsx
@@ -6,22 +6,19 @@ import { EModalAddressBookRoutes } from './types';
 import type { IModalAddressBookParamList } from './types';
 
 const AddressBookListModal = LazyLoadPage(
-  () =>
-    import('@onekeyhq/kit/src/common/components/AddressBook/pages/ListItem'),
+  () => import('@onekeyhq/kit/src/views/AddressBook/pages/ListItem'),
 );
 
 const AddressBookAddItemModal = LazyLoadPage(
-  () => import('@onekeyhq/kit/src/common/components/AddressBook/pages/AddItem'),
+  () => import('@onekeyhq/kit/src/views/AddressBook/pages/AddItem'),
 );
 
 const AddressBookEditItemModal = LazyLoadPage(
-  () =>
-    import('@onekeyhq/kit/src/common/components/AddressBook/pages/EditItem'),
+  () => import('@onekeyhq/kit/src/views/AddressBook/pages/EditItem'),
 );
 
 const AddressBookPickItemModal = LazyLoadPage(
-  () =>
-    import('@onekeyhq/kit/src/common/components/AddressBook/pages/PickItem'),
+  () => import('@onekeyhq/kit/src/views/AddressBook/pages/PickItem'),
 );
 
 export const ModalAddressBookRouter: IModalFlowNavigatorConfig<
